+++ tests/integration/func_sets/test_pricing.py
@@ -18,7 +18,7 @@ from envoy_schema.server.schema.sep2.pricing import (
 )
 from httpx import AsyncClient
 
-from envoy.server.mapper.sep2.mrid import PricingReadingType
+from envoy.server.mapper.constants import PricingReadingType
 from envoy.server.model.tariff import PRICE_DECIMAL_PLACES
 from tests.data.certificates.certificate1 import TEST_CERTIFICATE_FINGERPRINT as AGG_1_VALID_CERT
 from tests.integration.integration_server import cert_header
