+++ tests/integration/func_sets/test_response.py
@@ -20,7 +20,8 @@ from envoy_schema.server.schema.sep2.response import (
 from httpx import AsyncClient
 from sqlalchemy import func, insert, select
 
-from envoy.server.mapper.sep2.mrid import MridMapper, PricingReadingType, ResponseSetType
+from envoy.server.mapper.constants import PricingReadingType, ResponseSetType
+from envoy.server.mapper.sep2.mrid import MridMapper
 from envoy.server.mapper.sep2.response import response_set_type_to_href
 from envoy.server.model.doe import DynamicOperatingEnvelope
 from envoy.server.model.response import DynamicOperatingEnvelopeResponse, TariffGeneratedRateResponse
